src/requirements.txt
alembic==1.9.4
Babel==2.9.0
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
distlib==0.3.6
ecdsa==0.18.0
filelock==3.9.0
Flask==2.2.3
Flask-Cors==3.0.10
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-Moment==0.11.0
Flask-Script==2.0.6
Flask-SQLAlchemy==3.0.3
Flask-WTF==0.14.3
greenlet==2.0.2
gunicorn==20.1.0
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
markup==0.2
MarkupSafe==2.1.2
platformdirs==3.0.0
postgres==4.0
protobuf==4.21.12
psycopg2-binary==2.9.5
psycopg2-pool==1.1
pyasn1==0.4.8
pycparser==2.21
pyOpenSSL==23.1.1
python-dateutil==2.6.0
python-dotenv==0.21.0
python-jose==3.3.0
pytz==2022.7.1
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.4
typed-ast==1.5.4
typing_extensions==4.5.0
virtualenv==20.20.0
Werkzeug==2.2.3
wrapt==1.11.1
WTForms==3.0.1
zipp==3.15.0
